refactor: keep Tone decoding behind audio boundary
